Adds a thread
In this example you demonstrate how a property can be bound to an
element in a ObservableMap. However, the data in the map does not
change, and this is what this change does, i.e. a thread is started
which increases each sec the values by 10%.

This is a change in a part of the code which is not contained in the
book.