Changes pseudo.css
The code defines the used customRegion to be of type "my", therefore
this type should be used in the css file. Otherweise, the line
        customRegion.getStyleClass().add("my");
could be deleted.

As an alternative, both declarations could be shown by defining

   .my:active{
       -fx-background-color: red;
       -fx-opacity: 1;
   }

   CustomRegion:active{
       -fx-background-color: blue;
       -fx-opacity: 1;
   }

If the "my" style is not assigned, then the color changes to blue.